Module Orx.Config

val set_basename : string -> unit

val load : string -> Status.t

val load_from_memory : string -> Status.t

val set_bootstrap : (unit -> Status.t) -> unit

val push_section : string -> unit

val pop_section : unit -> unit

val get_current_section : unit -> string

val select_section : string -> unit

val get_section_count : unit -> int

val get_section : int -> string

val get_key_count : unit -> int

val get_key : int -> string

val get_parent : string -> string option

val has_section : string -> bool

val has_value : string -> bool

val clear_section : string -> Status.t

val clear_value : string -> Status.t

val get_string : string -> string

val set_string : string -> string -> unit

val get_bool : string -> bool

val set_bool : string -> bool -> unit

val get_float : string -> float

val set_float : string -> float -> unit

val get_int : string -> int

val set_int : string -> int -> unit

val get_vector : string -> Vector.t

val set_vector : string -> Vector.t -> unit

val get_list_vector : string -> int option -> Vector.t

val set_list_string : string -> string list -> unit

val append_list_string : string -> string list -> unit

val if_has_value : string -> (string -> 'a) -> 'a option

if_has_value key getter is Some (getter key) if key exists in the currently selected config section or None if key does not exist in the current section.

val exists : section:string -> key:string -> bool

exists ~section ~key is true if key exists in section.

val get : (string -> 'a) -> section:string -> key:string -> 'a

val set : (string -> 'a -> unit) -> 'a -> section:string -> key:string -> unit

val get_seq : (string -> 'a) -> section:string -> key:string -> 'a Stdlib.Seq.t

get_seq getter ~section ~key is a sequence of values pulled repeatedly from the same section and key.

If the values are random then a new random value will be returned for every element of the sequence.

If the section and key represent a constant value then the sequence will return the same value for every element.

If section and key do not exist then the result will be Seq.empty.

val get_list_item : (string -> int option -> 'a) -> int option -> section:string -> key:string -> 'a

val get_list : (string -> int option -> 'a) -> section:string -> key:string -> 'a list

val is_list : string -> bool

val get_sections : unit -> string list

val get_current_section_keys : unit -> string list

val get_section_keys : string -> string list

val get_guid : string -> Structure.Guid.t

val set_guid : string -> Structure.Guid.t -> unit

val with_section : string -> (unit -> 'a) -> 'a